backend/kfc_absolute/requirements.txt
celery==5.2.7
Django==4.2.1
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
environ==1.0
psycopg2==2.9.6
redis==4.5.5
